Add socket.io
Have app.js create the http server so it can also create io object which
is added to res.io in app.use so it's available to other operations and
finally return both app and server in module.export.

Change bin/www to get http server created in app.js.